refactor: cmdutil -> internal.cmdutil
